await sound before redirecting, but with a timeout
if the sound hasn't played in 1s, give up and just redirect
