+++ ui/analyse/src/explorer/explorerCtrl.ts
@@ -3,7 +3,7 @@ import { storedBooleanProp } from 'common/storage';
 import { defer } from 'common/defer';
 import { fenColor } from 'common/mini-board';
 import debounce from 'common/debounce';
-import { sync, Sync } from 'common/sync';
+import { sync, Sync } from 'common/promise';
 import { opposite } from 'chessground/util';
 import * as xhr from './explorerXhr';
 import { winnerOf } from './explorerUtil';
