--- ui/common/src/sync.ts
@@ -1,15 +1,0 @@
-export interface Sync<T> {
-  promise: Promise<T>;
-  sync: T | undefined;
-}
-
-export function sync<T>(promise: Promise<T>): Sync<T> {
-  const sync: Sync<T> = {
-    sync: undefined,
-    promise: promise.then(v => {
-      sync.sync = v;
-      return v;
-    }),
-  };
-  return sync;
-}
